UI update > Folder refactoring
