+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:character_wiki/Character_Wiki/Presentation/Bloc/Character/character_bloc.dart';
-import 'package:character_wiki/Character_Wiki/Domain/Character/home.dart';
+import 'package:character_wiki/Character_Wiki/Presentation/Home/home.dart';
 import 'package:character_wiki/service_locator.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
